extracted Product Option's CRUD to repository
