+++ refactor-me/refactor-me.csproj
@@ -98,6 +98,7 @@
     <Compile Include="Models\Helpers.cs" />
     <Compile Include="Models\Product.cs" />
     <Compile Include="Models\ProductOption.cs" />
+    <Compile Include="Models\ProductOptionRepository.cs" />
     <Compile Include="Models\ProductOptions.cs" />
     <Compile Include="Models\Products.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
